\name{lmorigin-package}
\alias{lmorigin-package}
\docType{package}
\title{ Multiple regression through the origin }
\description{
Function \code{\link{lmorigin}} computes a multiple linear regression and performs tests of significance of the equation parameters using permutations. The regression line can be forced through the origin. Testing the significance in that case requires a special permutation procedure. This option was developed for the analysis of independent contrasts, which requires regression through the origin.
}
\details{
\tabular{ll}{
Package: \tab lmorigin\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-22\cr
License: \tab Copyleft\cr
LazyLoad: \tab yes\cr
}
Function \code{\link{lmorigin}} computes a multiple linear regression and performs tests of significance of the equation parameters using permutations. The regression line can be forced through the origin.

}

\author{ Pierre Legendre }

\references{
Legendre, P. & Y. Desdevises. 2009. Independent contrasts and regression through the origin. Journal of Theoretical Biology (in press), doi:10.1016/j.jtbi.2009.04.022
}

\keyword{ package }
